+++ libfreeipmi/include/freeipmi/util/ipmi-util.h
@@ -29,17 +29,17 @@ extern "C" {
 
 uint8_t ipmi_checksum (uint8_t *buf, uint64_t len);
 
-int8_t ipmi_check_cmd (fiid_obj_t obj_cmd, uint8_t cmd);
+int ipmi_check_cmd (fiid_obj_t obj_cmd, uint8_t cmd);
 
-int8_t ipmi_check_completion_code (fiid_obj_t obj_cmd, uint8_t completion_code);
+int ipmi_check_completion_code (fiid_obj_t obj_cmd, uint8_t completion_code);
 
-int8_t ipmi_check_completion_code_success (fiid_obj_t obj_cmd);
+int ipmi_check_completion_code_success (fiid_obj_t obj_cmd);
 
 int ipmi_get_random (uint8_t *buf, unsigned int buflen);
 
-int8_t ipmi_is_ipmi_1_5_packet (uint8_t *pkt, uint32_t pkt_len);
+int ipmi_is_ipmi_1_5_packet (uint8_t *pkt, uint32_t pkt_len);
 
-int8_t ipmi_is_ipmi_2_0_packet (uint8_t *pkt, uint32_t pkt_len);
+int ipmi_is_ipmi_2_0_packet (uint8_t *pkt, uint32_t pkt_len);
 
 const char *ipmi_cmd_str (uint8_t net_fn, uint8_t cmd);
