+++ libfreeipmi/include/freeipmi/util/rmcp-util.h
@@ -28,7 +28,7 @@ extern "C" {
 #include <stdint.h>
 #include <freeipmi/fiid/fiid.h>
 
-int8_t ipmi_rmcp_check_message_tag (fiid_obj_t pong, uint8_t message_tag);
+int ipmi_rmcp_check_message_tag (fiid_obj_t pong, uint8_t message_tag);
 
 #ifdef __cplusplus
 }
